pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.2.0",
    "python-telegram-bot>=22.7",
    "qrcode[pil]>=8.2",
]
